+++ packages/react-core/src/filters/viewportFeaturesBinary.js
@@ -114,6 +114,10 @@ function calculateViewportFeatures({
   type,
   uniqueIdProperty
 }) {
+  if (!data?.properties.length) {
+    return;
+  }
+
   if (tileIsFullyVisible) {
     addAllFeaturesInTile({ map, data, uniqueIdProperty });
   } else {
@@ -158,8 +162,9 @@ export function viewportFeaturesBinary({ tiles, viewport, uniqueIdProperty }) {
   const map = new Map();
 
   for (const tile of tiles) {
-    // Discard if it's not a visible tile or tile has not data
-    if (!tile.isVisible || !tile.data) {
+    // Discard if it's not a visible tile (only check false value, not undefined)
+    // or tile has not data
+    if (tile.isVisible === false || !tile.data) {
       continue;
     }